v0.2.1: feat: Introduce AudioConverter utility and refactor audio_library
This commit introduces a new utility script 'audio_converter.py' and performs substantial refactoring on 'audio_library.py'. It also modifies the '__init__.py' file to incorporate the new class into the 'voxstruct.utils' package.

The 'audio_converter.py', added to 'src/voxstruct/utils', is a new feature that accommodates a class 'AudioConverter'. This class provides methods for converting and segmenting audio files and can be utilized both from the command line and as an imported module. The implementation includes options to customize output directories, segmentation parameters, and ability to handle multiple files simultaneously. This new class enhances the audio conversion functionality and makes it more accessible and easier to use across the project.

In the '__init__.py' file located in 'src/voxstruct/utils', the 'AudioConverter' has been imported and added to the '__all__' list. This change enhances the import functionality of the 'AudioConverter' class when the 'voxstruct.utils' package is imported, adhering to the principle of making every component easily accessible.

The 'audio_library.py' file in 'src/voxstruct/utils' has been significantly refactored. The refactor involves the removal of a large chunk of code that was previously used to determine the project root and import the 'AudioConverter' module. This code has been deemed unnecessary due to the inclusion of 'AudioConverter' in the 'voxstruct.utils' package. The refactoring simplifies the 'audio_library.py' file and removes a potentially error-prone method of importing the 'AudioConverter' module.

The purpose of these changes is to enhance audio conversion functionality and make it more accessible across the project. These changes adhere to the project's architectural pattern of making components easily accessible and enhance the code's simplicity by removal of complex and potentially error-prone code. However, these changes may impact other parts of the system, particularly those that depend on audio conversion, necessitating additional testing and possible updates.

Files Changed:
[M] src/voxstruct/utils/__init__.py
[A] src/voxstruct/utils/audio_converter.py
[M] src/voxstruct/utils/audio_library.py